Move report abstract to common abstract in prep for action abstract
